pyproject.toml
[tool.poetry]
name = "voice-activity"
version = "0.1.0"
description = ""
authors = ["Rafał Galczak <r.galczak@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
python-dotenv = "^0.15.0"
stdlib-list = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
